+++ src/error.rs
@@ -3,9 +3,9 @@
 use std::io;
 
 #[cfg(feature = "security")]
-use openssl::ssl::{self, Error as SslError};
-#[cfg(feature = "security")]
 use openssl::error::ErrorStack;
+#[cfg(feature = "security")]
+use openssl::ssl::{self, Error as SslError};
 
 /// The various errors this library can produce.
 error_chain! {
@@ -13,8 +13,6 @@ error_chain! {
         Io(io::Error) #[doc="Input/Output error while communicating with Kafka"];
 
         Ssl(SslError) #[cfg(feature = "security")] #[doc="An error as reported by OpenSsl"];
-
-        SslHandshake(ErrorStack) #[cfg(feature = "security")] #[doc="An error as reported by OpenSsl handshake"];
 
         InvalidSnappy(::snap::Error) #[cfg(feature = "snappy")] #[doc="Failure to encode/decode a snappy compressed response from Kafka"];
     }
@@ -201,9 +199,10 @@ pub enum KafkaCode {
 impl<S> From<ssl::HandshakeError<S>> for Error {
     fn from(err: ssl::HandshakeError<S>) -> Error {
         match err {
-            ssl::HandshakeError::SetupFailure(e) => From::from(e),
-            ssl::HandshakeError::Failure(s) |
-            ssl::HandshakeError::Interrupted(s) => from_sslerror_ref(s.error()).into(),
+            ssl::HandshakeError::SetupFailure(e) => from_sslerror_ref(&From::from(e)).into(),
+            ssl::HandshakeError::Failure(s) | ssl::HandshakeError::WouldBlock(s) => {
+                from_sslerror_ref(s.error()).into()
+            }
         }
     }
 }
@@ -218,8 +217,6 @@ impl Clone for Error {
             }
             #[cfg(feature = "security")]
             &Error(ErrorKind::Ssl(ref x), _) => from_sslerror_ref(x).into(),
-            #[cfg(feature = "security")]
-            &Error(ErrorKind::SslHandshake(ref x), _) => ErrorKind::SslHandshake(x.clone()).into(),
             &Error(ErrorKind::UnsupportedProtocol, _) => ErrorKind::UnsupportedProtocol.into(),
             &Error(ErrorKind::UnsupportedCompression, _) => {
                 ErrorKind::UnsupportedCompression.into()
@@ -239,91 +236,69 @@ impl Clone for Error {
 
 #[cfg(feature = "security")]
 fn from_sslerror_ref(err: &ssl::Error) -> ErrorKind {
-    match err {
-        &SslError::ZeroReturn => ErrorKind::Ssl(SslError::ZeroReturn),
-        &SslError::WantRead(ref e) => ErrorKind::Ssl(SslError::WantRead(clone_ioe(e))),
-        &SslError::WantWrite(ref e) => ErrorKind::Ssl(SslError::WantWrite(clone_ioe(e))),
-        &SslError::WantX509Lookup => ErrorKind::Ssl(SslError::WantX509Lookup),
-        &SslError::Stream(ref e) => ErrorKind::Ssl(SslError::Stream(clone_ioe(e))),
-        &SslError::Ssl(ref es) => ErrorKind::Ssl(SslError::Ssl(es.clone())),
+    if let Some(io_err) = err.io_error() {
+        return ErrorKind::Io(clone_ioe(io_err));
     }
+    if let Some(ssl_err) = err.ssl_error() {
+        return ErrorKind::Ssl(From::from(ssl_err.to_owned()));
+    }
+
+    // xxx: return something as default?
+    unreachable!()
 }
 
 #[cfg(feature = "snappy")]
 fn from_snap_error_ref(err: &::snap::Error) -> ErrorKind {
     match err {
         &::snap::Error::TooBig { given, max } => {
-            ErrorKind::InvalidSnappy(::snap::Error::TooBig {
-                given: given,
-                max: max,
-            })
+            ErrorKind::InvalidSnappy(::snap::Error::TooBig { given, max })
         }
         &::snap::Error::BufferTooSmall { given, min } => {
-            ErrorKind::InvalidSnappy(::snap::Error::BufferTooSmall {
-                given: given,
-                min: min,
-            })
+            ErrorKind::InvalidSnappy(::snap::Error::BufferTooSmall { given, min })
         }
         &::snap::Error::Empty => ErrorKind::InvalidSnappy(::snap::Error::Empty),
         &::snap::Error::Header => ErrorKind::InvalidSnappy(::snap::Error::Header),
         &::snap::Error::HeaderMismatch {
             expected_len,
             got_len,
-        } => {
-            ErrorKind::InvalidSnappy(::snap::Error::HeaderMismatch {
-                expected_len: expected_len,
-                got_len: got_len,
-            })
-        }
+        } => ErrorKind::InvalidSnappy(::snap::Error::HeaderMismatch {
+            expected_len,
+            got_len,
+        }),
         &::snap::Error::Literal {
             len,
             src_len,
             dst_len,
-        } => {
-            ErrorKind::InvalidSnappy(::snap::Error::Literal {
-                len: len,
-                src_len: src_len,
-                dst_len: dst_len,
-            })
-        }
+        } => ErrorKind::InvalidSnappy(::snap::Error::Literal {
+            len,
+            src_len,
+            dst_len,
+        }),
         &::snap::Error::CopyRead { len, src_len } => {
-            ErrorKind::InvalidSnappy(::snap::Error::CopyRead {
-                len: len,
-                src_len: src_len,
-            })
+            ErrorKind::InvalidSnappy(::snap::Error::CopyRead { len, src_len })
         }
         &::snap::Error::CopyWrite { len, dst_len } => {
-            ErrorKind::InvalidSnappy(::snap::Error::CopyWrite {
-                len: len,
-                dst_len: dst_len,
-            })
+            ErrorKind::InvalidSnappy(::snap::Error::CopyWrite { len, dst_len })
         }
         &::snap::Error::Offset { offset, dst_pos } => {
-            ErrorKind::InvalidSnappy(::snap::Error::Offset {
-                offset: offset,
-                dst_pos: dst_pos,
-            })
+            ErrorKind::InvalidSnappy(::snap::Error::Offset { offset, dst_pos })
         }
         &::snap::Error::StreamHeader { byte } => {
-            ErrorKind::InvalidSnappy(::snap::Error::StreamHeader { byte: byte })
+            ErrorKind::InvalidSnappy(::snap::Error::StreamHeader { byte })
         }
         &::snap::Error::StreamHeaderMismatch { ref bytes } => {
-            ErrorKind::InvalidSnappy(::snap::Error::StreamHeaderMismatch { bytes: bytes.clone() })
+            ErrorKind::InvalidSnappy(::snap::Error::StreamHeaderMismatch {
+                bytes: bytes.clone(),
+            })
         }
         &::snap::Error::UnsupportedChunkType { byte } => {
-            ErrorKind::InvalidSnappy(::snap::Error::UnsupportedChunkType { byte: byte })
+            ErrorKind::InvalidSnappy(::snap::Error::UnsupportedChunkType { byte })
         }
         &::snap::Error::UnsupportedChunkLength { len, header } => {
-            ErrorKind::InvalidSnappy(::snap::Error::UnsupportedChunkLength {
-                len: len,
-                header: header,
-            })
+            ErrorKind::InvalidSnappy(::snap::Error::UnsupportedChunkLength { len, header })
         }
         &::snap::Error::Checksum { expected, got } => {
-            ErrorKind::InvalidSnappy(::snap::Error::Checksum {
-                expected: expected,
-                got: got,
-            })
+            ErrorKind::InvalidSnappy(::snap::Error::Checksum { expected, got })
         }
     }
 }