+++ tests/integration/consumer_producer/producer.rs
@@ -1,4 +1,6 @@
 use super::*;
+
+use env_logger;
 use kafka::error;
 use kafka::producer::Record;
 
@@ -14,6 +16,7 @@ fn test_producer_send() {
 /// Sending to a non-existent topic should fail.
 #[test]
 fn test_producer_send_non_existent_topic() {
+    let _ = env_logger::init();
     let mut producer = test_producer();
 
     let error_code = match producer