+++ tests/kafka-rust-image/download-kafka.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+set -ex
+
+url="https://archive.apache.org/dist/kafka/${KAFKA_VERSION}/kafka_${SCALA_VERSION}-${KAFKA_VERSION}.tgz"
+curl -Sso "/tmp/kafka_${SCALA_VERSION}-${KAFKA_VERSION}.tgz" "${url}"